DRY Downloader.
