+++ activesupport/test/multibyte_grapheme_break_conformance_test.rb
@@ -1,29 +1,14 @@
 # encoding: utf-8
 
 require 'abstract_unit'
+require 'multibyte_test_helpers'
 
 require 'fileutils'
 require 'open-uri'
 require 'tmpdir'
 
 class MultibyteGraphemeBreakConformanceTest < ActiveSupport::TestCase
-  class Downloader
-    def self.download(from, to)
-      unless File.exist?(to)
-        $stderr.puts "Downloading #{from} to #{to}"
-        unless File.exist?(File.dirname(to))
-          system "mkdir -p #{File.dirname(to)}"
-        end
-        open(from) do |source|
-          File.open(to, 'w') do |target|
-            source.each_line do |l|
-              target.write l
-            end
-          end
-        end
-      end
-    end
-  end
+  include MultibyteTestHelpers
 
   TEST_DATA_URL = "http://www.unicode.org/Public/#{ActiveSupport::Multibyte::Unicode::UNICODE_VERSION}/ucd/auxiliary"
   TEST_DATA_FILE = '/GraphemeBreakTest.txt'
