+++ activesupport/test/multibyte_test_helpers.rb
@@ -1,4 +1,22 @@
 module MultibyteTestHelpers
+  class Downloader
+    def self.download(from, to)
+      unless File.exist?(to)
+        unless File.exist?(File.dirname(to))
+          system "mkdir -p #{File.dirname(to)}"
+        end
+        open(from) do |source|
+          File.open(to, 'w') do |target|
+            source.each_line do |l|
+              target.write l
+            end
+          end
+        end
+      end
+      true
+    end
+  end
+
   UNICODE_STRING = 'こにちわ'.freeze
   ASCII_STRING = 'ohayo'.freeze
   BYTE_STRING = "\270\236\010\210\245".force_encoding("ASCII-8BIT").freeze
